+++ components/cel/celfile.h
@@ -22,9 +22,9 @@ namespace Cel
             CelFile(const std::string& filename);
 
             //If normal cel file, returns same as numFrames(), for an archive, the number of frames in each subcel
-            size_t animLength() const;
-            size_t numFrames() const;
-            CelFrame& operator[] (size_t index);
+            int32_t animLength() const;
+            int32_t numFrames() const;
+            CelFrame& operator[] (int32_t index);
 
         private:
             CelDecoder mDecoder;
