+++ components/cel/celframe.cpp
@@ -2,12 +2,12 @@
 
 namespace Cel
 {
-    const Colour& get(size_t x, size_t y, const CelFrame& frame)
+    const Colour& get(int32_t x, int32_t y, const CelFrame& frame)
     {
         return frame.mRawImage[x + (frame.mHeight-1-y)*frame.mWidth];
     }
      
-    Misc::Helper2D<const CelFrame, const Colour&> CelFrame::operator[] (size_t x) const
+    Misc::Helper2D<const CelFrame, const Colour&> CelFrame::operator[] (int32_t x) const
     {
         return Misc::Helper2D<const CelFrame, const Colour&>(*this, x, get);
     }
