+++ components/cel/celframe.h
@@ -14,15 +14,15 @@ namespace Cel
     class CelFrame
     {
         public:
-            size_t mWidth;
-            size_t mHeight;
+            int32_t mWidth;
+            int32_t mHeight;
             
-            Misc::Helper2D<const CelFrame, const Colour&> operator[] (size_t x) const;
+            Misc::Helper2D<const CelFrame, const Colour&> operator[] (int32_t x) const;
             
         private:
             friend class CelFile;
             friend class CelDecoder;
-            friend const Colour& get(size_t x, size_t y, const CelFrame& frame);
+            friend const Colour& get(int32_t x, int32_t y, const CelFrame& frame);
             
             std::vector<Colour> mRawImage;
     };
