+++ components/level/level.cpp
@@ -5,8 +5,8 @@
 namespace Level
 {
     Level::Level(const Dun& dun, const std::string& tilPath, const std::string& minPath,
-        const std::string& solPath, const std::string& tileSetPath, const std::pair<size_t,size_t>& downStairs,
-        const std::pair<size_t,size_t>& upStairs, std::map<size_t, size_t> doorMap, size_t previous, size_t next):
+        const std::string& solPath, const std::string& tileSetPath, const std::pair<int32_t,int32_t>& downStairs,
+        const std::pair<int32_t,int32_t>& upStairs, std::map<int32_t, int32_t> doorMap, int32_t previous, int32_t next):
         mDun(dun), mTil(tilPath), mMin(minPath), mSol(solPath), mTileSetPath(tileSetPath),
 
         mTilPath(tilPath), mMinPath(minPath), mSolPath(solPath), mDoorMap(doorMap), mUpStairs(upStairs), mDownStairs(downStairs),
@@ -14,7 +14,7 @@ namespace Level
 
     std::vector<int16_t> Level::mEmpty(16);
 
-    bool Level::isStairs(size_t x, size_t y) const
+    bool Level::isStairs(int32_t x, int32_t y) const
     {
         if(mDownStairs.first == x && mDownStairs.second == y)
             return true;
@@ -24,10 +24,10 @@ namespace Level
         return false;
     }
 
-    const MinPillar get(size_t x, size_t y, const Level& level)
+    const MinPillar get(int32_t x, int32_t y, const Level& level)
     {
-        size_t xDunIndex = x;
-        size_t xTilIndex = 0;
+        int32_t xDunIndex = x;
+        int32_t xTilIndex = 0;
         if((xDunIndex % 2) != 0)
         {
             xDunIndex--;
@@ -35,8 +35,8 @@ namespace Level
         }
         xDunIndex /= 2;
 
-        size_t yDunIndex = y;
-        size_t yTilIndex = 0;
+        int32_t yDunIndex = y;
+        int32_t yTilIndex = 0;
         if((yDunIndex % 2) != 0)
         {
             yDunIndex--;
@@ -44,7 +44,7 @@ namespace Level
         }
         yDunIndex /= 2;
 
-        size_t tilIndex;
+        int32_t tilIndex;
         
         if(xTilIndex)
         {
@@ -66,36 +66,36 @@ namespace Level
         if(dunIndex == -1)
             return MinPillar(Level::mEmpty, 0, -1);
 
-        size_t minIndex = level.mTil[dunIndex][tilIndex];
+        int32_t minIndex = level.mTil[dunIndex][tilIndex];
         
         return MinPillar(level.mMin[minIndex], level.mSol.passable(minIndex), minIndex);
     }
 
-    Misc::Helper2D<const Level, const MinPillar> Level::operator[] (size_t x) const
+    Misc::Helper2D<const Level, const MinPillar> Level::operator[] (int32_t x) const
     {
         return Misc::Helper2D<const Level, const MinPillar>(*this, x, get);
     }
             
-    void Level::activate(size_t x, size_t y)
+    void Level::activate(int32_t x, int32_t y)
     {
-        size_t xDunIndex = x;
+        int32_t xDunIndex = x;
         if((xDunIndex % 2) != 0)
             xDunIndex--;
         xDunIndex /= 2;
 
-        size_t yDunIndex = y;
+        int32_t yDunIndex = y;
         if((yDunIndex % 2) != 0)
             yDunIndex--;
         yDunIndex /= 2;
 
-        size_t index = mDun[xDunIndex][yDunIndex]; 
+        int32_t index = mDun[xDunIndex][yDunIndex]; 
         
         // open doors when clicked on
         if(mDoorMap.find(index) != mDoorMap.end())
             mDun[xDunIndex][yDunIndex] = mDoorMap[index];
     }
     
-    size_t Level::minSize() const
+    int32_t Level::minSize() const
     {
         return mMin.size();
     }
@@ -105,22 +105,22 @@ namespace Level
         return MinPillar(mMin[i], mSol.passable(i), i);
     }
 
-    size_t Level::width() const
+    int32_t Level::width() const
     {
         return mDun.width()*2;
     }
 
-    size_t Level::height() const
+    int32_t Level::height() const
     {
         return mDun.height()*2;
     }
 
-    const std::pair<size_t,size_t>& Level::upStairsPos() const
+    const std::pair<int32_t,int32_t>& Level::upStairsPos() const
     {
         return mUpStairs;
     }
 
-    const std::pair<size_t,size_t>& Level::downStairsPos() const
+    const std::pair<int32_t,int32_t>& Level::downStairsPos() const
     {
         return mDownStairs;
     }
@@ -137,12 +137,12 @@ namespace Level
 
     MinPillar::MinPillar(const std::vector<int16_t>& data, bool passable, int32_t index): mData(data), mPassable(passable), mIndex(index) {}
 
-    size_t MinPillar::size() const
+    int32_t MinPillar::size() const
     {
         return mData.size();
     }
 
-    int16_t MinPillar::operator[] (size_t index) const
+    int16_t MinPillar::operator[] (int32_t index) const
     {
         return mData[index];
     }
