+++ components/level/level.h
@@ -23,8 +23,8 @@ namespace Level
     class MinPillar
     {
         public:
-            size_t size() const;
-            int16_t operator[] (size_t index) const;
+            int32_t size() const;
+            int16_t operator[] (int32_t index) const;
             bool passable() const;
             int32_t index() const;
 
@@ -36,7 +36,7 @@ namespace Level
             int32_t mIndex;
 
             friend class Level;
-            friend const MinPillar get(size_t x, size_t y, const Level& level);
+            friend const MinPillar get(int32_t x, int32_t y, const Level& level);
     };
 
 
@@ -44,28 +44,28 @@ namespace Level
     {
         public:
             Level(const Dun& dun, const std::string& tilPath, const std::string& minPath,
-                const std::string& solPath, const std::string& tileSetPath, const std::pair<size_t,size_t>& downStairs,
-                const std::pair<size_t,size_t>& upStairs, std::map<size_t, size_t> doorMap, size_t previous, size_t next);
+                const std::string& solPath, const std::string& tileSetPath, const std::pair<int32_t,int32_t>& downStairs,
+                const std::pair<int32_t,int32_t>& upStairs, std::map<int32_t, int32_t> doorMap, int32_t previous, int32_t next);
 
             Level() {}
 
-            Misc::Helper2D<const Level, const MinPillar> operator[] (size_t x) const;
+            Misc::Helper2D<const Level, const MinPillar> operator[] (int32_t x) const;
 
-            void activate(size_t x, size_t y);
+            void activate(int32_t x, int32_t y);
 
-            size_t minSize() const;
+            int32_t minSize() const;
             const MinPillar minPillar(int32_t i) const;
 
-            size_t width() const;
-            size_t height() const;
+            int32_t width() const;
+            int32_t height() const;
 
-            const std::pair<size_t,size_t>& upStairsPos() const;
-            const std::pair<size_t,size_t>& downStairsPos() const;
+            const std::pair<int32_t,int32_t>& upStairsPos() const;
+            const std::pair<int32_t,int32_t>& downStairsPos() const;
 
             const std::string& getTileSetPath() const;
             const std::string& getMinPath() const;
 
-            bool isStairs(size_t, size_t) const;
+            bool isStairs(int32_t, int32_t) const;
 
             int32_t getNextLevel() const
             {
@@ -87,13 +87,13 @@ namespace Level
             std::string mMinPath; ///< path to min file for level
             std::string mSolPath; ///< path to sol file for this level
 
-            std::map<size_t, size_t> mDoorMap; ///< Map from closed door indices to open door indices + vice-versa
+            std::map<int32_t, int32_t> mDoorMap; ///< Map from closed door indices to open door indices + vice-versa
 
-            std::pair<size_t,size_t> mUpStairs;
-            std::pair<size_t,size_t> mDownStairs;
+            std::pair<int32_t,int32_t> mUpStairs;
+            std::pair<int32_t,int32_t> mDownStairs;
 
             static std::vector<int16_t> mEmpty;
-            friend const MinPillar get(size_t x, size_t y, const Level& level);
+            friend const MinPillar get(int32_t x, int32_t y, const Level& level);
 
             int32_t mPrevious; ///< index of previous level
             int32_t mNext; ///< index of next level
